Changed again to use unsigned int in Graph flags for link indexes, allowing up to 64 outgoing links.
git-svn-id: svn+ssh://ruttval.infra.kth.se/svn/mezzo@313 be8bedfb-b93a-dc11-ad5b-00609709f61e

+++ trunk/mezzo_lib/src/Graph.cpp
@@ -48,8 +48,8 @@ template <class T, class I>
 void Graph<T, I>::addLink(const int i, const int u,const  int d, // required
 						  const T w,	// following is optional
 						  const char grade,
-						  const unsigned short int legal,
-						  const unsigned short int index)
+						  const unsigned int legal,
+						  const unsigned int index)
 {
    GraphLink<T, I> *arc = link(i);
    arc->upNode_ = u;
